enh(viz): add gestures to scene 3D
These changes will allow to use gestures on the 3D scene. It allows in particular to pinch fingers to zoom and to long tap to place a landmark. Simple taps and pan gestures should intuitively map to mouse clicks and mouse drags.

+++ modules/viz/scene3d/adaptor/SShapeExtruder.cpp
@@ -362,7 +362,7 @@ std::tuple<Ogre::Vector3, Ogre::Vector3, Ogre::Vector3> SShapeExtruder::getNearF
 
 //-----------------------------------------------------------------------------
 
-void SShapeExtruder::wheelEvent(Modifier /*_mods*/, int /*_angleDelta*/, int /*_x*/, int /*_y*/)
+void SShapeExtruder::wheelEvent(Modifier /*_mods*/, double /*_angleDelta*/, int /*_x*/, int /*_y*/)
 {
     if(m_interactionEnableState)
     {
